requirements.txt
flask
psutil
scikit-learn
pandas
numpy
mysql-connector-pythonpython-dotenv
flask-compress